randomize start index with a fixed seed
if the start index is the same in case of multiple threads, we could get
cached results. In order to prevent that, we should start from a random
index in case we are using multiple threads. However, when this
benchmark is run across different platforms, we should be using the same
random start indexes in the same order to be fair. Thefore seed is also
supplied to the random function. Ideally someone could generate a random
seed, and use the same seed when running the benchmark across different platforms.