Allow easily rerunning of download-hammerdb.sh
If the download of the HammerDB release failed for some reason you'd
have to remove the tarball to try downloading it again. Otherwise it
would not redownload and only check the known incorrect hash of the file
that's already there.

This downloads to a temp file and only renames it to the final tarball
name once the hash check completes successfully.