Show NOPM in stdout at the end of the benchmark run
To find the NOPM after a run you would have to either scroll up to see
it in the output or you'd have to manually look at the contents of the
`hammerdb_nopm_xxx.log` file. This changes our script so that the NOPM
is shown as the final line of output of the benchmark run.